requirements.txt
streamlit
openai
python-pptx
pypdf